changed ntupelwriter_94X to ntuplewriter and changed xml-template naminge sceme in create-multi-dataset-xmlfiles
