+++ scripts/gc-conf/DATA_2017_example.conf
@@ -20,7 +20,7 @@ se path = srm://dcache-se-cms.desy.de:8443/srm/managerv2?SFN=/pnfs/desy.de/cms/t
 [CMSSW]
 project area = /nfs/dust/cms/user/garbersc/UHH2_17/CMSSW_9_4_1/
 software requirements    = False
-config file = /nfs/dust/cms/user/garbersc/UHH2_17/CMSSW_9_4_1/src/UHH2/core/python/ntuplewriter_94X.py
+config file = /nfs/dust/cms/user/garbersc/UHH2_17/CMSSW_9_4_1/src/UHH2/core/python/ntuplewriter.py
 se runtime      = True 
 dataset provider = DBS3Provider
 events per job = 10000
